renamed direcotry and adjusted namespaces, started with operator tests
