+++ src/Sulu/Bundle/ResourceBundle/Resource/Exception/FilterDependencyNotFoundException.php
@@ -7,30 +7,36 @@
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
  */
-namespace Sulu\Bundle\ResourceBundle\Filter\Exception;
+
+namespace Sulu\Bundle\ResourceBundle\Resource\Exception;
 
 /**
- * Class FilterNotFoundException
+ * This exception is thrown if a required dependency for creating a new
+ * filter is not given.
  */
-class FilterNotFoundException extends FilterException
+class FilterDependencyNotFoundException extends FilterException
 {
     /**
      * The name of the object not found
      * @var string
      */
     private $entityName;
-
     /**
      * The id of the object not found
      * @var integer
      */
     private $id;
 
-    public function __construct($id)
+    public function __construct($entityName, $id)
     {
-        $this->entityName = 'SuluResourceBundle:Filter';
+        $this->entityName = $entityName;
         $this->id = $id;
-        parent::__construct('The filter with the id "'.$this->id.'" was not found.', 0);
+        parent::__construct(
+            'The filter dependency "'.$this->entityName.
+            ' with the id "'.$this->id.
+            '" was not found.',
+            0
+        );
     }
 
     /**
@@ -50,5 +56,4 @@ class FilterNotFoundException extends FilterException
     {
         return $this->id;
     }
-
 }
