+++ src/Sulu/Bundle/ResourceBundle/Resource/Exception/FilterNotFoundException.php
@@ -7,36 +7,30 @@
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
  */
-
-namespace Sulu\Bundle\ResourceBundle\Filter\Exception;
+namespace Sulu\Bundle\ResourceBundle\Resource\Exception;
 
 /**
- * This exception is thrown if a required dependency for creating a new
- * filter is not given.
+ * Class FilterNotFoundException
  */
-class FilterDependencyNotFoundException extends FilterException
+class FilterNotFoundException extends FilterException
 {
     /**
      * The name of the object not found
      * @var string
      */
     private $entityName;
+
     /**
      * The id of the object not found
      * @var integer
      */
     private $id;
 
-    public function __construct($entityName, $id)
+    public function __construct($id)
     {
-        $this->entityName = $entityName;
+        $this->entityName = 'SuluResourceBundle:Filter';
         $this->id = $id;
-        parent::__construct(
-            'The filter dependency "'.$this->entityName.
-            ' with the id "'.$this->id.
-            '" was not found.',
-            0
-        );
+        parent::__construct('The filter with the id "'.$this->id.'" was not found.', 0);
     }
 
     /**
@@ -56,4 +50,5 @@ class FilterDependencyNotFoundException extends FilterException
     {
         return $this->id;
     }
+
 }
