+++ src/Sulu/Bundle/ResourceBundle/Resource/Exception/MissingConditionAttributeException.php
@@ -8,7 +8,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ResourceBundle\Filter\Exception;
+namespace Sulu\Bundle\ResourceBundle\Resource\Exception;
 
 /**
  * This exception is thrown if a required property for creating or manipulating
